+++ packages/codec/index.ts
@@ -1,6 +1,7 @@
 export * from "./decoder";
 export * from "./encoder";
 export * from "./descriptors";
+export { View } from "./view";
 
 // TODO [ToDr] Temporary re-export bytes until we have proper publishes.
 export * as bytes from "@typeberry/bytes";
